register client

+++ src/models/client.rs
@@ -1,5 +1,6 @@
+#[derive(Default)]
 pub struct Client{
-    pub id: i32,
+    pub id: usize,
     pub name: String,
     pub cpf: String,
     pub address: String
